// models/job/job-detail-model.ts
//   const JobDetailModel = JobDetailBuilder.toJobDetailModel(json);

export interface JobDetailModel {
    jobTypeString:       string;
    salaryTypeString:    string;
    id:                  string;
    inReason:            string;
    enReason:            string;
    jobId:               string;
    jobName:             string;
    jobDetail:           string;
    jobDuration:         number;
    jobSalary:           string;
    jobMaxSalary:        string;
    jobDate:             string;
    jobTime:             string;
    jobBased:            string;
    jobType:             string;
    jobTag:              string;
    jobAddress:          string;
    jobNote:             string;
    jobLat:              string;
    jobLng:              string;
    jobAge:              string;
    jobDistance:         number;
    commisionSalary:     string;
    jobCity:             string;
    jobTypingAddress:    string;
    jobProvider:         string;
    providerName:        string;
    jobTypeId:           number;
    minAge:              string;
    maxAge:              string;
    scam:                number;
    scamStat:            number;
    jobGender:           string;
    bosVerified:         number;
    bosId:               string;
    bosEmail:            string;
    bosPhone:            string;
    status:              string;
    publish:             string;
    updated_at:          string;
    created_at:          Date;
    view:                number;
    dtSort:              Date;
    categoryId:          string;
    categoryString:      string;
    categoryChild:       string;
    categoryChildString: string;
    applied:             string;
    jobsPhoto:           string;
    salaryTypeId:        number;
    salaryPeriodeId:     number;
    salaryNote:          string;
    minEducationId:      string;
    bosName:             string;
    bosPhoto:            string;
    salaryPeriodeString: string;
    minEducationString:  string;
    jobProviderString:   string;
    categoryLogo:        string;
    reason:              string;
    jobBasedString:      string;
    jobCityString:       string;
}